backend/github_service/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.11
djangorestframework==3.15.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.9
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.3.0
